refactor(agent): fold @agent/storage deep-imports behind the storage barrel
#10011 established module-level barrels as the fold-in proving ground; this
applies that pattern to the @agent/storage cluster.

- Extend src/agent/storage/index.ts with the five remaining symbols CLI still
  reached for by deep path (formatConversationMessage, completeOwnedExecutionLease,
  ExecutionLeaseLostError, markOwnedExecutionLeaseUndurable, OwnedExecutionLeaseScope);
  inspectExecutionLease was already exported.
- Migrate all five host import sites onto @agent/storage.
- Drop the two migrated specifiers from host-agent-import-baseline.json
  (cli 17 -> 15); no baseline is widened.
- Reconcile the -08-04 review §5.1/§6 wording with the intermediate step #10011
  actually took (module-level barrels as the proving ground, packages/agent
  promotion deferred), citing review pullrequestreview-4918028384.

Refs #10024

Remaining clusters for the issue: followUp, core/definition/AgentConfig.
