+++ packages/cli/src/runtime/history/conversationFormat.ts
@@ -1,4 +1,4 @@
-import { formatConversationMessage } from '@agent/storage/conversationFormat';
+import { formatConversationMessage } from '@agent/storage';
 
 import type {
   CliHistoryConversationPreview,
